Refactor account registration to use CPR instead of UUID
- Updated the registration feature to reflect that users receive a CPR upon successful registration instead of a UUID.
- Modified the AccountCreated and AccountDeregistered events to include CPR instead of accountId.
- Adjusted the Account aggregate to remove accountId and use CPR for identification.
- Changed ReadAccountRepository methods to use CPR for account retrieval and deregistration.
- Updated WriteAccountRepository to save and retrieve accounts using CPR.
- Refactored AccountService methods to accept and return CPR instead of accountId.
- Adjusted test steps and context to handle CPR instead of accountId.
- Updated DtuPayClient and PaymentService to work with CPR in payment processing.
- Refined PaymentRequest and PaymentRecord models to use BigDecimal for amounts.
- Enhanced error handling in TokenServiceClient for better timeout management.

+++ dtu-pay-server/src/test/java/dtu/pay/services/PaymentServiceTest.java
@@ -9,6 +9,7 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Consumer;
 
@@ -23,8 +24,10 @@ public class PaymentServiceTest {
         public void publish(Event event) {
             publishedEvent.complete(event);
         }
+
         @Override
-        public void addHandler(String eventType, Consumer<Event> handler) {}
+        public void addHandler(String eventType, Consumer<Event> handler) {
+        }
     };
 
     @BeforeEach
@@ -40,7 +43,7 @@ public class PaymentServiceTest {
 
     @Test
     void registerSuccessfully() throws Exception {
-        PaymentRequest paymentReq = new PaymentRequest("random_token_here", "merchantId_here", 1000);
+        PaymentRequest paymentReq = new PaymentRequest("random_token_here", "merchantId_here", BigDecimal.valueOf(1000));
         publishedEvent = new CompletableFuture<>();
 
         CompletableFuture<String> resultFuture = CompletableFuture.supplyAsync(() -> {
@@ -59,8 +62,7 @@ public class PaymentServiceTest {
         String expectedResponse = "Success!";
         paymentService.handlePaymentSuccessful(new Event(
                 "PaymentSuccessful",
-                new Object[] { expectedResponse, correlationId }
-        ));
+                new Object[] { expectedResponse, correlationId }));
 
         assertEquals(expectedResponse, resultFuture.get());
     }
